requirements.txt
Flask==3.0.3
google-cloud-speech==2.26.0
openai==1.40.3
httpx==0.27.*         
python-dotenv==1.0.1